+++ app/src/main/java/com/yikangcheng/admin/yikang/bean/RecommendBean.java
@@ -43,8 +43,8 @@ public class RecommendBean {
     private int id;
     private String name;
     private int isPay;
-    private int sourceprice;
-    private int currentprice;
+    private double sourceprice;
+    private double currentprice;
     private String title;
     private String logo;
     private int losetype;
@@ -74,19 +74,19 @@ public class RecommendBean {
         this.isPay = isPay;
     }
 
-    public int getSourceprice() {
+    public double getSourceprice() {
         return sourceprice;
     }
 
-    public void setSourceprice(int sourceprice) {
+    public void setSourceprice(double sourceprice) {
         this.sourceprice = sourceprice;
     }
 
-    public int getCurrentprice() {
+    public double getCurrentprice() {
         return currentprice;
     }
 
-    public void setCurrentprice(int currentprice) {
+    public void setCurrentprice(double currentprice) {
         this.currentprice = currentprice;
     }
